Calender  user wise issue done
2. Calender function not working
"When a User Accept an Appointment the entrie should added in the calender. But at the moment it added to the calender who is logged in. When in the Appointment responsive User is "manurkoc" but at this time Email was accepted from admin and admin was at this time logged in the entrie added in the calender of the admin. THATS WRONG !!!. it doesn't matter who click on accept in the email. It should be added in the calender of the responive user who was choosen in the entrie. "

Try by yourself. 1. Send email to lead@dk-..... 2. Than start cron. 3. Than logged out from xics. 3. Than click in the Appoinment Email on "Accept". Normally the date must be added in the calender of the Responsive User which is in the entrie. But no Calender entrie there???