New Menu Callcenter 10-11-2020 done
Change this Icon on button. See S1.png
Use a Mail icon.

After click this button Leadstatus changed to "Terminbestätigung verschickt". When this status is set
so change background color of this row to "Yellow" color.

Check all Apointment Emails. See S2.png. There must stand as example Dateart: Webmeeting and not
a Date. Check O2, Telekom and Vodafoneemails.

Changings on Landingoages. See S3.png

The logo must be have a small distance to the top.
Write this line in bold font
There stand "TerminTermin" >> change to "Termin"
I click in the Email on "Accept" button and the Event was added in the Calender of mansurkoc.
Normally must be all entries which was added in the calender in red color. Bute here is in
green color. It must be everytime in the color of the responsive user.

I click in the calender on the event and than the edit button. There i see that the
adres of the Lead was added in the Inputfield "Event Start Adress". But it must be added in
the Inputfield "Event Location". See S4.png

See S5.png. When i click in the Deatilview of one entrie under "Terminierung" on the button "Lead erstellen"
in the green Popup there Stand "Send Mail Successfully". Thats wrong. There should stand
"Lead was added" in english lang and "Lead wurde hinzugefügt" in german lang.