termination live changes done
1. On Landingpages on the red button stand "Accept" >> "Zusagen"

2. In last email to lead@dk.... I have choosed by field:
added by: mansur koc

But it has added the entrie with admin.?!

3. After I click on Accept in Appointment Email the entrie was not added in the calender. See last entrie.

4. Please add a new User role called "Agent". He was only able to see
the Menu "Tickets" and only his entries and he was able to see Menu "Terminierung" and only his entries. And under Terminierung he was not able to see the button "Export". He can't see the button to send appoinment email. He cant delete entries from Terminierung. ANd in detailview of an entrie he cant see the button "Lead erstellen"

5. The Button "Export" can only see admin and SalesManager

6. In the appointmentemails from Telekom and Vodafone cant see the logos?